# Add workaround for Firefox video play performance issue