@dezigns333 2021-05-01
The best video performance I have had was with the new Microsoft Edge
browser.
60 FPS even while screen sharing a video from a tab with either Youtube or
Netflix.
I was able to use VB-Audio virtual cable as a loopback to send the audio
through the mic so it was received perfectly in Hubs.

On Sat, May 1, 2021 at 1:00 AM Brian Peiris ***@***.***>
wrote:

> ***@***.**** approved this pull request.
>
> My tests on various devices wiith a 4K video:
> RGB RGBA
> FF on Windows 6fps 60fps
> FF on Mac 20fps 60fps
> FF on Linux 14fps 14fps
> FxR on Quest 35fps 35fps
>
> Firefox on Android: Could not test due to existing bug where videos don't
> render at all :( #4213 <https://github.com/mozilla/hubs/issues/4213>
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/mozilla/hubs/pull/4210#pullrequestreview-649727851>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AATDKS7BG3QJMKUMYYF64SLTLODFZANCNFSM435IQSCA>
> .
>