@dezigns333 2021-05-03
It works even better when Edge is capturing video from Firefox. Avoid
Chrome.

On Mon, May 3, 2021 at 12:06 PM Takahiro ***@***.***> wrote:

> Merged #4210 <https://github.com/mozilla/hubs/pull/4210> into master.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/mozilla/hubs/pull/4210#event-4678929915>, or
> unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AATDKSZXGBZK267GA5FEGPLTL3CZDANCNFSM435IQSCA>
> .
>